requirements.txt
streamlit
jieba
wordcloud
Pillow
numpy
